Skill ut IM expandable innhold
